requirements.txt
Pillow
python-twitter
instapy-cli
python-telegram-bot
requests